new fixes

+++ include/my_types.hh
@@ -5,11 +5,11 @@
 #include <boost/cstdint.hpp>
 #include <boost/integer_traits.hpp>
 
-typedef int32_t   node_t;
-typedef int64_t   edge_t;
-typedef float     cost_t;
-typedef int32_t   resource_t;
-typedef float     dist_t;
+typedef int16_t   node_t;
+typedef int32_t   edge_t;
+typedef double    cost_t;
+typedef double    resource_t;
+typedef double    dist_t;
 
 #include <vector>
 typedef std::vector<resource_t> resources;
